refactor(admin): remove unused auth code and consolidate hooks
Delete unused admin shell, session guard, site placeholder page, and duplicate access-token hook. Merge admin session into useAdminUser, drop redundant page-level auth checks, and centralize Supabase config helpers.
