+++ src/app/admin/(dashboard)/layout.tsx
@@ -1,5 +1,3 @@
-import { AdminSessionGuard } from "@/components/admin/admin-session-guard";
-import { isAdminAuthBypassEnabled } from "@/lib/auth/admin-bypass";
 import { requireAdminSession } from "@/lib/auth/require-admin-session";
 
 export const dynamic = "force-dynamic";
@@ -10,10 +8,5 @@ export default async function AdminDashboardLayout({
   children: React.ReactNode;
 }>) {
   await requireAdminSession();
-
-  if (isAdminAuthBypassEnabled()) {
-    return <AdminSessionGuard enabled={false}>{children}</AdminSessionGuard>;
-  }
-
-  return <AdminSessionGuard enabled>{children}</AdminSessionGuard>;
+  return children;
 }
