+++ src/app/admin/forgot-password/reset/page.tsx
@@ -16,7 +16,7 @@ import {
 import { formatAuthError } from "@/lib/auth/format-auth-error";
 import { checkPassword } from "@/lib/auth/password";
 import { clearRecoveryEmail } from "@/lib/auth/recovery-storage";
-import { isSupabaseBrowserConfigured } from "@/lib/auth/supabase-configured";
+import { isSupabaseBrowserConfigured } from "@/lib/supabase/public-config";
 import { createSupabaseBrowserClient } from "@/lib/supabase/browser";
 
 export default function ForgotPasswordResetPage() {
