+++ src/app/admin/forgot-password/verify/page.tsx
@@ -17,7 +17,7 @@ import { formatAuthError } from "@/lib/auth/format-auth-error";
 import { getPasswordRecoveryRedirectUrl } from "@/lib/auth/recovery-redirect";
 import { maskEmail } from "@/lib/auth/mask-email";
 import { setRecoveryEmail } from "@/lib/auth/recovery-storage";
-import { isSupabaseBrowserConfigured } from "@/lib/auth/supabase-configured";
+import { isSupabaseBrowserConfigured } from "@/lib/supabase/public-config";
 import { createSupabaseBrowserClient } from "@/lib/supabase/browser";
 
 export default function ForgotPasswordVerifyPage() {
