--- src/lib/auth/supabase-configured.ts
@@ -1,5 +1,0 @@
-import { getSupabaseAnonKey, getSupabaseUrl } from "@/lib/supabase/public-config";
-
-export function isSupabaseBrowserConfigured(): boolean {
-  return Boolean(getSupabaseUrl() && getSupabaseAnonKey());
-}
